requirements.txt
aiohttp
discord.py
python-dotenv
websockets
asyncpg